sp-41: remove scroll from app-content, refactor
