+++ src/components/templates/CentralBlock.vue
