+++ src/views/Login.vue
@@ -1,8 +1,5 @@
 <template>
-  <CenteredBlockTemplate :messages="pagesMessages['login']">
-    <template #title>
-      {{ lang.title.login }}
-    </template>
+  <CentralBlock :messages="pagesMessages['login']">
     <form class="login-page__form">
       <UniversalField
         :label="lang.label.email"
@@ -62,7 +59,7 @@
         />
       </div>
     </template>
-  </CenteredBlockTemplate>
+  </CentralBlock>
 </template>
 <script lang="ts" setup>
 import { onBeforeMount, ref } from "vue";
@@ -80,7 +77,7 @@ import { useAppStore } from "@/store/appStore";
 import UniversalButton from "@/components/buttons/UniversalButton.vue";
 import { useRoute, useRouter } from "vue-router";
 import { storeToRefs } from "pinia";
-import CenteredBlockTemplate from "@/components/templates/CenteredBlockTemplate.vue";
+import CentralBlock from "@/components/templates/CentralBlock.vue";
 import { getValidationErrors } from "@/helpers/formValidation";
 import UniversalField from "@/components/fields/UniversalField.vue";
 import type { ApiValidationError } from "@/types/common";
